Renamed file for SimulatedNmeaDataSource to match new classname
