+++ lib/samples/show_device_location_with_nmea_data_sources/README.metadata.json
@@ -26,8 +26,8 @@
         "NmeaSatelliteInfo"
     ],
     "snippets": [
-        "nmea_source_simulator.dart",
-        "show_device_location_with_nmea_data_sources.dart"
+        "show_device_location_with_nmea_data_sources.dart",
+        "simulated_nmea_data_source.dart"
     ],
     "title": "Show device location with NMEA data sources"
 }
