+++ lib/samples/show_device_location_with_nmea_data_sources/show_device_location_with_nmea_data_sources.dart
@@ -17,7 +17,7 @@ import 'dart:convert';
 
 import 'package:arcgis_maps/arcgis_maps.dart';
 import 'package:arcgis_maps_sdk_flutter_samples/common/common.dart';
-import 'package:arcgis_maps_sdk_flutter_samples/samples/show_device_location_with_nmea_data_sources/nmea_source_simulator.dart';
+import 'package:arcgis_maps_sdk_flutter_samples/samples/show_device_location_with_nmea_data_sources/simulated_nmea_data_source.dart';
 import 'package:arcgis_maps_sdk_flutter_samples/utils/sample_state_support.dart';
 import 'package:flutter/material.dart';
