+++ lib/samples/show_device_location_with_nmea_data_sources/simulated_nmea_data_source.dart
